refactor parsing, add cli
